refactor(project): move shelflist loading to new usePlaylists hook
